Consistently use GCP XML API
